Separated this from other API module
There is still a lot of work to be done but I got rid of the old module.
It was stupid to rely on software config to distinguish between local and
non local users. I'd rather just choose what port to listen on or what
machine is actually listening. So I got rid of the old stuff and added
some new configuration values. Needs better error handling because right
now it just breaks if the subdomain does not exist.